mini-server-prod/kgroot-services/ai-sre-platform/requirements.txt
# AI SRE Platform Dependencies

# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# Neo4j
neo4j==5.14.0

# OpenAI (for GPT-5)
openai==1.3.5

# Async
aiohttp==3.9.1
asyncio==3.4.3

# Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
ruff==0.1.6
mypy==1.7.1

# Monitoring (optional)
prometheus-client==0.19.0
